+++ src/main/java/fi/csc/chipster/auth/AuthenticationService.java
@@ -18,15 +18,15 @@ import fi.csc.chipster.auth.jaas.JaasAuthenticationProvider;
 import fi.csc.chipster.auth.model.OidcLoginSession;
 import fi.csc.chipster.auth.model.Role;
 import fi.csc.chipster.auth.model.User;
+import fi.csc.chipster.auth.oidc.OidcProvidersImpl;
+import fi.csc.chipster.auth.oidc.loginsessions.OidcLoginSessions;
+import fi.csc.chipster.auth.oidc.loginsessions.OidcLoginSessionsInDb;
+import fi.csc.chipster.auth.oidc.loginsessions.OidcLoginSessionsInMemory;
 import fi.csc.chipster.auth.resource.AuthAdminResource;
 import fi.csc.chipster.auth.resource.AuthTokenResource;
 import fi.csc.chipster.auth.resource.AuthTokens;
 import fi.csc.chipster.auth.resource.AuthUserResource;
 import fi.csc.chipster.auth.resource.AuthenticationRequestFilter;
-import fi.csc.chipster.auth.resource.OidcLoginSessions;
-import fi.csc.chipster.auth.resource.OidcLoginSessionsInDb;
-import fi.csc.chipster.auth.resource.OidcLoginSessionsInMemory;
-import fi.csc.chipster.auth.resource.OidcProvidersImpl;
 import fi.csc.chipster.auth.resource.OidcResource;
 import fi.csc.chipster.auth.resource.UserTable;
 import fi.csc.chipster.rest.Config;
