+++ src/main/java/fi/csc/chipster/auth/oidc/OidcConfig.java
@@ -1,4 +1,4 @@
-package fi.csc.chipster.auth.resource;
+package fi.csc.chipster.auth.oidc;
 
 public class OidcConfig {
