+++ src/main/java/fi/csc/chipster/auth/resource/OidcResource.java
@@ -31,6 +31,10 @@ import com.nimbusds.openid.connect.sdk.validators.IDTokenValidator;
 import fi.csc.chipster.auth.model.OidcLoginSession;
 import fi.csc.chipster.auth.model.Role;
 import fi.csc.chipster.auth.model.UserId;
+import fi.csc.chipster.auth.oidc.NimbusHelpers;
+import fi.csc.chipster.auth.oidc.OidcConfig;
+import fi.csc.chipster.auth.oidc.OidcProviders;
+import fi.csc.chipster.auth.oidc.loginsessions.OidcLoginSessions;
 import fi.csc.chipster.rest.Config;
 import fi.csc.chipster.rest.RestUtils;
 import fi.csc.chipster.rest.hibernate.Transaction;
